refactor: Reorganize packages
