+++ src/main/java/gift/controller/page/AdminController.java
@@ -1,4 +1,4 @@
-package gift.controller;
+package gift.controller.page;
 
 import gift.entity.ProductRecord;
 import gift.repository.ProductDAO;
