+++ src/main/java/gift/dto/user/EncryptedUpdateDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.user;
 
 public record EncryptedUpdateDTO(
         long id,
