+++ src/main/java/gift/dto/user/TokenResponseDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.user;
 
 public record TokenResponseDTO(String token) {
