+++ src/main/java/gift/dto/wish/WishCreateDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.wish;
 
 public record WishCreateDTO(
         long userId,
