+++ src/main/java/gift/dto/wish/WishInfoDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.wish;
 
 public record WishInfoDTO(
         long id,
