+++ src/main/java/gift/dto/wish/WishRequestDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.wish;
 
 public record WishRequestDTO(
         long productId,
