+++ src/main/java/gift/dto/wish/WishResponseDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.wish;
 
 public record WishResponseDTO(
         long id,
