+++ src/main/java/gift/repository/WishDAO.java
@@ -1,7 +1,7 @@
 package gift.repository;
 
-import gift.dto.WishCreateDTO;
-import gift.dto.WishInfoDTO;
+import gift.dto.wish.WishCreateDTO;
+import gift.dto.wish.WishInfoDTO;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
